Conflicts:
	galera/src/ist.hpp
	galera/src/replicator.hpp
	galera/tests/ist_check.cpp
	gcs/src/gcs.hpp

+++ gcache/src/gcache.h
@@ -21,9 +21,9 @@ typedef struct _gcache gcache_t;
 extern gcache_t* gcache_create  (gu_config_t* conf, const char* data_dir);
 extern void      gcache_destroy (gcache_t* gc);
 
-extern void* gcache_malloc      (gcache_t* gc, size_t size);
+extern void* gcache_malloc      (gcache_t* gc, int size);
 extern void  gcache_free        (gcache_t* gc, const void* ptr);
-extern void* gcache_realloc     (gcache_t* gc, void* ptr, size_t size);
+extern void* gcache_realloc     (gcache_t* gc, void* ptr, int size);
 
 extern int64_t gcache_seqno_min (gcache_t* gc);
